requirements.txt
# Core AI dependencies
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
langchain-classic==0.3.13

# UI
streamlit==1.31.0
rich==13.9.4

# Google Cloud / BigQuery
google-cloud-bigquery==3.28.0
pyarrow==17.0.0
pandas==2.2.3
google-cloud-secret-manager==2.21.1

# Configuration
python-dotenv==1.0.1